third_party/blink/web_tests/http/tests/htmlimports/import-script-block-crossorigin-dynamic-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
This test ensures that crossorigin-marked script elements are blocked properly

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.externalScriptHasRun is true
PASS document.corsExternalScriptHasRun is true
PASS document.corsExternalScriptForLocalhostHasRun is true
PASS successfullyParsed is true

TEST COMPLETE
